+++ app/src/main/java/com/luismibm/trackify/MainLayout.kt
@@ -1,0 +1,46 @@
+package com.luismibm.trackify
+
+import androidx.compose.material3.DrawerState
+import androidx.compose.material3.DrawerValue
+import androidx.compose.material3.ModalDrawerSheet
+import androidx.compose.material3.ModalNavigationDrawer
+import androidx.compose.material3.NavigationDrawerItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import kotlinx.coroutines.launch
+
+@Composable
+fun MainLayout(navController: NavHostController, userEmail: String) {
+
+    val drawerState = remember { DrawerState(initialValue = DrawerValue.Closed) }
+    val coroutineScope = rememberCoroutineScope()
+
+    ModalNavigationDrawer(
+        drawerState = drawerState,
+        drawerContent = {
+            ModalDrawerSheet {
+                NavigationDrawerItem(
+                    label = { Text(text = "Home") },
+                    selected = false,
+                    onClick = {
+                        coroutineScope.launch {
+                            drawerState.close()
+                        }
+                        navController.navigate("home")
+                    }
+                )
+            }
+        }
+    ) {
+        NavHost(navController = navController, startDestination = "home") {
+            composable("home") { MainScreen(navController = navController, email = userEmail) }
+        }
+    }
+
+}
