+++ app/src/main/java/com/luismibm/trackify/auth/AuthManager.kt
@@ -1,0 +1,53 @@
+package com.luismibm.trackify.auth
+
+import android.content.Context
+import com.luismibm.trackify.BuildConfig
+import io.github.jan.supabase.auth.Auth
+import io.github.jan.supabase.auth.auth
+import io.github.jan.supabase.auth.providers.builtin.Email
+import io.github.jan.supabase.createSupabaseClient
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+
+sealed interface AuthResponse {
+    data object Success: AuthResponse
+    data class Error(val message: String?): AuthResponse
+}
+
+class AuthManager(
+    private val context: Context
+) {
+
+    private val supabase = createSupabaseClient(
+        supabaseUrl = BuildConfig.SUPABASE_URL,
+        supabaseKey = BuildConfig.SUPABASE_KEY
+    ) {
+        install(Auth)
+    }
+
+    fun signUp(emailValue: String, passwordValue: String): Flow<AuthResponse> = flow {
+        try {
+            supabase.auth.signUpWith(Email) {
+                email = emailValue
+                password = passwordValue
+            }
+            emit(AuthResponse.Success)
+        } catch (e: Exception) {
+            emit(AuthResponse.Error(e.localizedMessage))
+        }
+    }
+
+    fun signIn(emailValue: String, passwordValue: String): Flow<AuthResponse> = flow {
+        try {
+            supabase.auth.signInWith(Email) {
+                email = emailValue
+                password = passwordValue
+            }
+            emit(AuthResponse.Success)
+        } catch (e: Exception) {
+            emit(AuthResponse.Error(e.localizedMessage))
+        }
+
+    }
+
+}
